+++ package.json
@@ -46,6 +46,7 @@
 		"bits-ui": "^0.21.13",
 		"clsx": "^2.1.1",
 		"embla-carousel-svelte": "^8.1.7",
+		"idb-keyval": "^6.2.1",
 		"ionic-svelte": "^0.5.84",
 		"lucide-svelte": "^0.416.0",
 		"svelte-carousel": "^1.0.25",
