+++ src/lib/components/ui/label/index.ts
@@ -1,0 +1,7 @@
+import Root from "./label.svelte";
+
+export {
+	Root,
+	//
+	Root as Label,
+};
