+++ src/lib/tabStore.ts
@@ -1,0 +1,7 @@
+import { writable } from 'svelte/store';
+
+export const activeTab = writable<string>('');
+
+export function changeTab(tabName: string) {
+    activeTab.set(tabName);
+}
